+++ src/components/UI/Tabs/Tabs.js
@@ -1,29 +1,10 @@
+import PropTypes from 'prop-types';
 import React, {Fragment} from 'react';
-import {useLocation} from 'react-router-dom';
 
-import {useTabsTranslation} from '../../../hooks';
-import {useApp} from '../../../providers/AppProvider';
 import {Divider} from '../Divider/Divider';
 import {Tab} from '../Tab/Tab';
 
-export const Tabs = () => {
-  const {navigateToRoute} = useApp();
-  const {termsTxt, faqTxt} = useTabsTranslation();
-  const {pathname} = useLocation();
-
-  const tabs = [
-    {
-      text: termsTxt,
-      isActive: pathname === '/terms',
-      onClick: () => navigateToRoute('/terms')
-    },
-    {
-      text: faqTxt,
-      isActive: pathname === '/faq',
-      onClick: () => navigateToRoute('/faq')
-    }
-  ];
-
+export const Tabs = ({tabs}) => {
   return tabs.map((tab, index) => {
     return (
       <Fragment key={index}>
@@ -32,4 +13,8 @@ export const Tabs = () => {
       </Fragment>
     );
   });
+};
+
+Tabs.propTypes = {
+  tabs: PropTypes.arrayOf(PropTypes.object)
 };
